mypage에서 history 부분 분리
