+++ HowsFront/src/pages/Home/History/Qna/Qna.jsx
