--- HowsFront/src/pages/Home/Main/Main.jsx
@@ -1,10 +1,0 @@
-import styles from './Main.module.css'
-import test from '../../../assets/images/테스트.jpg'
-
-export const Main = () => {
-    return (
-        <div>
-            <img src={test} alt="" />
-        </div>
-    )
-}
